Rename use cases
