+++ src/main/kotlin/com/projectcitybuild/plugin/commands/UnbanIPCommand.kt
@@ -2,7 +2,7 @@ package com.projectcitybuild.plugin.commands
 
 import com.projectcitybuild.core.utilities.Failure
 import com.projectcitybuild.core.utilities.Success
-import com.projectcitybuild.features.bans.usecases.UnbanIPUseCase
+import com.projectcitybuild.features.bans.usecases.UnbanIP
 import com.projectcitybuild.support.spigot.commands.InvalidCommandArgumentsException
 import com.projectcitybuild.support.spigot.commands.SpigotCommand
 import com.projectcitybuild.support.spigot.commands.SpigotCommandInput
@@ -10,7 +10,7 @@ import com.projectcitybuild.support.textcomponent.send
 import javax.inject.Inject
 
 class UnbanIPCommand @Inject constructor(
-    private val unbanIPUseCase: UnbanIPUseCase,
+    private val unbanIP: UnbanIP,
 ) : SpigotCommand {
 
     override val label = "unbanip"
@@ -24,7 +24,7 @@ class UnbanIPCommand @Inject constructor(
 
         val targetIP = input.args.first()
 
-        val result = unbanIPUseCase.unbanIP(
+        val result = unbanIP.unbanIP(
             ip = targetIP,
             unbannerUUID = input.player.uniqueId,
             unbannerName = input.player.name,
@@ -33,8 +33,8 @@ class UnbanIPCommand @Inject constructor(
         when (result) {
             is Failure -> input.sender.send().error(
                 when (result.reason) {
-                    UnbanIPUseCase.FailureReason.IP_NOT_BANNED -> "$targetIP is not currently banned"
-                    UnbanIPUseCase.FailureReason.INVALID_IP -> "$targetIP is not a valid IP"
+                    UnbanIP.FailureReason.IP_NOT_BANNED -> "$targetIP is not currently banned"
+                    UnbanIP.FailureReason.INVALID_IP -> "$targetIP is not a valid IP"
                 }
             )
             is Success -> input.sender.send().success("IP $targetIP has been unbanned")
