+++ src/test/com/projectcitybuild/features/bans/usecases/UnbanIPTest.kt
@@ -1,6 +1,6 @@
 package com.projectcitybuild.features.bans.usecases
 
-class UnbanIPUseCaseTest {
+class UnbanIPTest {
 
 //    private lateinit var useCase: UnbanIPUseCase
 //
