+++ src/test/com/projectcitybuild/features/bans/usecases/UnbanUUIDTest.kt
@@ -3,109 +3,87 @@ package com.projectcitybuild.features.bans.usecases
 import com.projectcitybuild.core.utilities.Failure
 import com.projectcitybuild.repositories.PlayerBanRepository
 import com.projectcitybuild.repositories.PlayerUUIDRepository
-import com.projectcitybuild.support.spigot.kick.PlayerKicker
 import kotlinx.coroutines.test.runTest
 import org.bukkit.Server
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.mockito.ArgumentMatchers.anyString
 import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
 import org.mockito.kotlin.any
-import org.mockito.kotlin.eq
 import org.powermock.api.mockito.PowerMockito.mock
 import org.powermock.api.mockito.PowerMockito.`when`
 import java.util.UUID
 
-class BanUUIDUseCaseTest {
+class UnbanUUIDTest {
 
-    private lateinit var useCase: BanUUIDUseCase
+    private lateinit var useCase: UnbanUUID
 
     private lateinit var playerBanRepository: PlayerBanRepository
     private lateinit var playerUUIDRepository: PlayerUUIDRepository
-    private lateinit var playerKicker: PlayerKicker
     private lateinit var server: Server
 
     @BeforeEach
     fun setUp() {
         playerBanRepository = mock(PlayerBanRepository::class.java)
         playerUUIDRepository = mock(PlayerUUIDRepository::class.java)
-        playerKicker = mock(PlayerKicker::class.java)
         server = mock(Server::class.java)
 
-        useCase = BanUUIDUseCase(
+        useCase = UnbanUUID(
             playerBanRepository,
             playerUUIDRepository,
             server,
-            playerKicker,
         )
     }
 
     @Test
-    fun `ban should fail when player doesn't exist`() = runTest {
+    fun `unban should fail when player doesn't exist`() = runTest {
         val playerName = "banned_player"
 
         `when`(playerUUIDRepository.get(playerName)).thenReturn(null)
 
-        val result = useCase.ban(playerName, null, "staff_player", null)
+        val result = useCase.unban(playerName, null)
 
-        assertEquals(result, Failure(BanUUIDUseCase.FailureReason.PlayerDoesNotExist))
+        assertEquals(result, Failure(UnbanUUID.FailureReason.PlayerDoesNotExist))
     }
 
     @Test
-    fun `ban should fail when player is already banned`() = runTest {
+    fun `unban should fail when player is not banned`() = runTest {
         val playerName = "banned_player"
         val playerUUID = UUID.randomUUID()
-        val staffName = "staff_player"
         val staffUUID = UUID.randomUUID()
 
         `when`(playerUUIDRepository.get(playerName)).thenReturn(playerUUID)
-        `when`(playerBanRepository.ban(playerUUID, playerName, staffUUID, staffName, null))
-            .thenThrow(PlayerBanRepository.PlayerAlreadyBannedException())
+        `when`(playerBanRepository.unban(playerUUID, staffUUID))
+            .thenThrow(PlayerBanRepository.PlayerNotBannedException())
 
-        val result = useCase.ban(playerName, staffUUID, staffName, null)
+        val result = useCase.unban(playerName, staffUUID)
 
-        assertEquals(result, Failure(BanUUIDUseCase.FailureReason.PlayerAlreadyBanned))
+        assertEquals(result, Failure(UnbanUUID.FailureReason.PlayerNotBanned))
     }
 
     @Test
-    fun `ban should pass input to ban repository`() = runTest {
+    fun `unban should pass input to ban repository`() = runTest {
         val playerName = "banned_player"
         val playerUUID = UUID.randomUUID()
         val staffUUID = UUID.randomUUID()
-        val staffName = "staff_player"
-        val reason = "reason"
 
         `when`(playerUUIDRepository.get(playerName)).thenReturn(playerUUID)
 
-        useCase.ban(playerName, staffUUID, "staff_player", reason)
+        useCase.unban(playerName, staffUUID)
 
         verify(playerBanRepository, times(1))
-            .ban(playerUUID, playerName, staffUUID, staffName, reason)
+            .unban(playerUUID, staffUUID)
     }
 
     @Test
-    fun `ban should be broadcasted to all online players`() = runTest {
+    fun `unban should be broadcasted to all online players`() = runTest {
         val playerName = "banned_player"
 
         `when`(playerUUIDRepository.get(playerName)).thenReturn(UUID.randomUUID())
 
-        useCase.ban(playerName, UUID.randomUUID(), "staff_player", "reason")
+        useCase.unban(playerName, UUID.randomUUID())
 
         verify(server).broadcastMessage(any())
-    }
-
-    @Test
-    fun `ban should kick the online player`() = runTest {
-        val playerName = "banned_player"
-        val playerUUID = UUID.randomUUID()
-
-        `when`(playerUUIDRepository.get(playerName)).thenReturn(playerUUID)
-
-        useCase.ban(playerName, UUID.randomUUID(), "staff_player", "reason")
-
-        verify(playerKicker, times(1))
-            .kickByUUID(eq(playerUUID), anyString(), eq(PlayerKicker.KickContext.FATAL))
     }
 }
